backend/requirements.txt
openai>=1.0.0
anthropic>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-anthropic>=0.1.0
sqlalchemy>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
pytest>=7.4.0
pytest-cov>=4.1.0